+++ CHANGELOG.md
@@ -96,6 +96,8 @@
 - GP2-315 - route to market backend integration
 
 ### Implemented enhancements
+- GP2-748 - Analytics product classifier - match found
+- GP2-689 - Analytics product classifier - product selected
 - GP2-690 - Analytics target market open
 - GP2-739 - Add beta banner to learning landing page
 - GP2-853 - Coming soon functionality for CTA displayed on the Export plan landing page